Fixed pictures and citations
Hi,

I've managed to fix the pictures and citation issues. The citations were fixed by adding the .bib file and finding the appropriate way to cite things in jupyter online (I think I got lucky because for some reason everything worked out for me). The images were a bit harder task to fix. The main problem here is that jupyter books doesn't really `recognize' things that aren't HTML files. So, what I had to do was to recreate the images as .HTML files and then display those HTML files. This is plot2, plot3, plot4 in the file. 

However, there are some additional problems that I have encountered. The first is that section four takes a really long time to run, and when I was creating the jupyter book it often terminates (running section four).  The other problem was this error when building the book:

`jupyter_client.kernelspec.NoSuchKernel: No such kernel named dell38'

I have no clue if this is a Samuel-only problem, but I can't seem to find the solution to this problem online. In desperation,  I ran the following line in the terminal

`conda install -n analysis_1 jupyter'

For some reason, this fixes the problem when it appears for section 4, but does not fix the problem when it appears for section 5. 

I will include the build file in a separate patch. 

Thanks,
Samuel Zhao